no abi change version change run

+++ include/abi-relevant-file.hpp
@@ -1,1 +1,1 @@
-inline long x = 5;
+inline int x = 5;
